test-requirements.txt
# Test Code Dependencies
distributed
pytz
cloudpickle
freezegun

# Test Framework
hypothesis
pytest>=4.5.0
pytest-cov
pytest-mock
setuptools-scm

# Benchmarks
asv

# Documentation
sphinx
sphinx_rtd_theme
sphinx-click
IPython

# Formatting / Checking
pre-commit
flake8-mutable
